@@ -1,13 +1,13 @@
-name: utf
-version: 0.9.0+5
+name: utf_convert
+version: 0.10.0
 
 description: Provides common operations for manipulating Unicode sequences
 author: Dart Team <misc@dartlang.org>
 homepage: https://www.github.com/dart-lang/utf
 
 environment:
-  sdk: '>=2.0.0-dev.20.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
-  test: ^1.2.0
-  pedantic: ^1.8.0
+  test: ^1.16.8
+  pedantic: ^1.11.0
