@@ -5,7 +5,7 @@
 library utf.utf_test;
 
 import 'package:test/test.dart';
-import 'package:utf/utf.dart';
+import 'package:utf_convert/utf.dart';
 
 import 'expect.dart' as expect;
 
